requirements.txt
django
djangorestframework
django-mathfilters
mysqlclient
